Classes carried out to Resources class Library. GlobalUsings added.
