+++ OKR_Steam/Business/IBS/IItemBusiness.cs
@@ -1,6 +1,4 @@
-﻿using OKR_Steam.Models.ResponseModels;
-
-namespace OKR_Steam.Business.IBS
+﻿namespace OKR_Steam.Business.IBS
 {
     public interface IItemBusiness
     {
