+++ OKR_Steam/Controllers/C/UserController.cs
@@ -1,9 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using OKR_Steam.Business.IBS;
-using OKR_Steam.Models.DBModels;
-using OKR_Steam.Models.DBModels.DBRequestModels;
-using OKR_Steam.Models.RequestModels;
-using OKR_Steam.Models.ResponseModels;
 using OKR_Steam.Validators.UserValidators;
 
 namespace OKR_Steam.Controllers.C
@@ -26,9 +22,6 @@ namespace OKR_Steam.Controllers.C
         public ProcessResult<SteamProfileModel> GetSteamProfileDataByName(string username)
         {
             var returnData = new ProcessResult<SteamProfileModel>();
-
-
-            throw new Exception("hata");
 
             var requestData = new SteamProfileRequestModel()
             {
