+++ OKR_Steam/DataAccess/DA/ErrorDataAccess.cs
@@ -1,6 +1,4 @@
 ﻿using OKR_Steam.DataAccess.IDA;
-using OKR_Steam.Models.DBModels.Tables;
-using OKR_Steam.Models.ResponseModels;
 
 namespace OKR_Steam.DataAccess.DA
 {
