+++ OKR_Steam/DataAccess/IDA/IItemDataAccess.cs
@@ -1,7 +1,4 @@
-﻿using OKR_Steam.Models.DBModels.DBResponseModels;
-using OKR_Steam.Models.ResponseModels;
-
-namespace OKR_Steam.DataAccess.IDA
+﻿namespace OKR_Steam.DataAccess.IDA
 {
     public interface IItemDataAccess
     {
