+++ OKR_Steam/Validators/UserValidators/SteamProfileRequestModelValidator.cs
@@ -1,5 +1,4 @@
 ﻿using FluentValidation;
-using OKR_Steam.Models.RequestModels;
 
 namespace OKR_Steam.Validators.UserValidators
 {
