+++ Resources/DBModels/ResponseModels/SteamProfileDBResponse.cs
@@ -1,4 +1,10 @@
-﻿namespace OKR_Steam.Models.DBModels.DBResponseModels
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Resources.DBModels.ResponseModels
 {
     public class SteamProfileDBResponse
     {
@@ -7,7 +13,7 @@
         public string Username { get; set; }
         public int ProfileState { get; set; }
         public string ProfileURL { get; set; }
-        public Guid PrimaryClanId  { get; set; }
+        public Guid PrimaryClanId { get; set; }
         public string TradeURL { get; set; }
         public DateTime LastUpdated { get; set; }
     }
