+++ Resources/DBModels/Tables/ErrorLogDatabaseModel.cs
@@ -1,7 +1,14 @@
-﻿namespace OKR_Steam.Models.ResponseModels
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Resources.DBModels.Tables
 {
-    public class ErrorModel
-    { 
+    public class ErrorLogDatabaseModel
+    {
+        public Guid Id { get; set; }
         public DateTime Date { get; set; }
         public bool IsAuthenticated { get; set; }
         public string ExceptionType { get; set; }
@@ -9,10 +16,4 @@
         public string Message { get; set; }
         public int StatusCode { get; set; }
     }
-     
-              
-                
-                
-                 
-               
 }
