+++ Resources/RequestModels/SteamProfileRequestModel.cs
@@ -1,8 +1,8 @@
-﻿namespace OKR_Steam.Models.RequestModels
+﻿namespace Resources.RequestModels
 {
     public class SteamProfileRequestModel
-    {        
-        public string Username { get; set; }       
+    {
+        public string Username { get; set; }
         public string ProfileURL { get; set; }
         public string TradeURL { get; set; }
